requirements.txt
groq
edge-tts
moviepy
requests
python-dotenv
